Update index.md
Fixed issues with the special arm-gcc installation instructions for Ubuntu 14.x.  The versions that are referenced are not available (not sure if they are just old or maybe incorrect due to typos) and the commands did not work for me.  I updated the references to the versions found at https://launchpad.net/~terry.guo/+archive/ubuntu/gcc-arm-embedded.  These work.