add vcs description to build info header
git describe is quite useful to get a revision info based on the vcs
which also honors tags. so let's put this info in a new define into the
build info header.
for git 'git describe --always --tags' is used, for mercurial a somewhat
equivalent found at [1] as mercurial itself doesn't provide a describe
subcommand.
doc has been updated accordingly.
the units test unly checks for a description string longer than 1

[1] http://stackoverflow.com/questions/6693209/is-there-an-equivalent-
to-gits-describe-function-for-mercurial

Signed-off-by: Martin Gysel <me@bearsh.org>